show notification if playback open fails with DVBLINK_REMOTE_STATUS_INVALID_DATA
This shows a notification with "Playback failed, please ensure you are using TV MOSAIC PLUS".
Has used on begin only the smaller without "PLUS" and was first confused where the error is.

Not sure that this error type only relate to them, but to have a message can help new
users about and not become a question mark like me.